Prim algorithm
It was implemented the same kruskal idea, using the mapping logic of the vertices for different forests , summarizing each vertex is in a different forest. However, this logic of each vertex in a different forest is not necessary for the prim, but it works. Ideally, add the vertex to a single forest each iteration.